requirements.txt
flask==2.0.1
flask-socketio==5.1.1
flask-login==0.5.0
SpeechRecognition==3.8.1
pyaudio==0.2.11
opencv-python==4.7.0.72
mediapipe==0.10.8
numpy==1.24.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.70.0
python-dotenv==0.19.0
sentence-transformers==2.2.2
pathlib==1.0.1
librosa==0.10.1
sounddevice==0.4.6
scipy>=1.7.0
scikit-learn
speechbrain
torchaudio
torch
pickle-mixin
flask-wtf==1.2.1
python-socketio==5.4.0
python-engineio==4.2.1 